Fix: JourneyGetRes dto fix

+++ journey-service/src/main/java/com/ssafy/journeymate/journeyservice/dto/response/JourneyGetRes.java
@@ -11,8 +11,8 @@ public class JourneyGetRes {
     private String title;
     private Integer day;
     private Integer sequence;
-    private float xcoordinate;
-    private float ycoordinate;
+    private double xcoordinate;
+    private double ycoordinate;
 
 
 }
